re-org to fix #17
